Permissions functions are now safely nested in try/except statements
In situations where permissions can't be changed (which is most
situations), exceptions will not be raised.

+++ python2.6libs/hou_asset_mgr.py
@@ -7,6 +7,7 @@ import os, glob
 import hou
 import subprocess
 from ui_tools import ui, messageSeverity, fileMode
+from miscutil import fileutil
 
 import utilities as amu #asset manager utilites
 
@@ -94,6 +95,7 @@ def copyToOtlDir(node, filename, newName, newDef):
     newfilepath = os.path.join(OTLDIR, filename)
     oldfilepath = os.path.join(USERDIR, filename)
     node.type().definition().copyToHDAFile(newfilepath, new_name=newName, new_menu_name=newDef)
+    fileutil.clobberPermissions(newfilepath)
     switchOPLibraries(oldfilepath, newfilepath)
 
 def moveToOtlDir(node, filename):
@@ -110,6 +112,7 @@ def copyToUsrDir(node, filename):
     newfilepath = os.path.join(USERDIR, filename)
     oldfilepath = os.path.join(OTLDIR, filename)
     node.type().definition().copyToHDAFile(newfilepath)
+    fileutil.clobberPermissions(newfilepath)
     switchOPLibraries(oldfilepath, newfilepath)
 
 def lockOTL(filename):
@@ -446,6 +449,7 @@ def newContainer(hpath):
             amu.createNewAssetFolders(ASSETSDIR, filename)
             templateNode.type().definition().copyToHDAFile(newfilepath, new_name=filename, new_menu_name=name)
             hou.hda.installFile(newfilepath, change_oplibraries_file=True)
+            fileutil.clobberPermissions(newfilepath)
             newnode = hou.node(hpath).createNode(filename)
         else:
             ui.infoWindow("Asset by that name already exists. Cannot create asset.", wtitle='Asset Name', msev=messageSeverity.Error)
@@ -732,8 +736,6 @@ def newTexture():
             args = ['txmake','-mode','periodic','-compression','zip']
             args += ['-format','openexr','-half',convertedTexture,finalTexture]
 
-            subprocess.check_call(args)
-
             # Uncomment the following and comment out the previous call if PRMan is not present
             """
             args = 'iconvert -d half ' + convertedTexture + ' ' 
@@ -741,22 +743,29 @@ def newTexture():
 
             subprocess.check_call( args.split() )
             """
-            
-            # Rename texture and move into production pipeline 
-            finalTextureName, ext = os.path.splitext(os.path.basename(finalTexture))
 
-            newTextureName = assetName + '_' + geoName + '_' + shadingPass + ext
- 
-            newfilepath = os.path.join(assetImageDir,newTextureName)
+            try:
+                subprocess.check_call(args)
+            except subprocess.CalledProcessError as e:
+                ui.infoWindow('Failed to convert texture. The following error occured:\n' + str(e))
+            else:
+               # Rename texture and move into production pipeline 
+                finalTextureName, ext = os.path.splitext(os.path.basename(finalTexture))
 
-            shutil.copy(finalTexture,newfilepath)
+                newTextureName = assetName + '_' + geoName + '_' + shadingPass + ext
+     
+                newfilepath = os.path.join(assetImageDir,newTextureName)
 
-            # Remove temporary files
-            os.remove(finalTexture)
-            os.remove(convertedTexture)
+                #shutil.copy(finalTexture,newfilepath)
+                fileutil.move(finalTexture, newfilepath)  
 
-            # Output final success message
-            ui.infoWindow('Your texture was saved to: ' + newfilepath + didgamma)
+                if convertedTexture != userTextureMap:
+                    os.remove(convertedTexture)
+
+                # Output final success message
+                ui.infoWindow('Your texture was saved to: ' + newfilepath + didgamma)
+            finally:
+                pass
 
 def getNodeInfo(node):
     if node != None and isDigitalAsset(node):
